[docs] static version selector in the Reference sidebar
