+++ docs/ui/components/Sidebar/ApiVersionSelect.tsx
@@ -1,0 +1,55 @@
+import { ChevronDownIcon } from '@expo/styleguide-icons';
+
+import { usePageApiVersion } from '~/providers/page-api-version';
+import versions from '~/public/static/constants/versions.json';
+import { A, FOOTNOTE, CALLOUT } from '~/ui/components/Text';
+
+const { VERSIONS, LATEST_VERSION, BETA_VERSION } = versions;
+
+// TODO(cedric): move this to a generic select input, so we can reuse it in the color scheme selector
+
+export function ApiVersionSelect() {
+  const { version, hasVersion, setVersion } = usePageApiVersion();
+
+  if (!hasVersion) {
+    return null;
+  }
+
+  return (
+    <div className="px-4 pt-3 pb-4 flex flex-col gap-1 border-b border-b-default bg-default">
+      <FOOTNOTE theme="tertiary">Reference version</FOOTNOTE>
+      <div className="relative bg-default border border-default rounded-md shadow-xs py-2 px-3">
+        <label className="flex flex-row items-center justify-between" htmlFor="api-version-select">
+          <CALLOUT className="flex">{versionToText(version)}</CALLOUT>
+          <ChevronDownIcon className="icon-sm shrink-0" />
+        </label>
+        <select
+          id="api-version-select"
+          className="absolute opacity-0 inset-0 w-full h-full text-xs cursor-pointer"
+          value={version}
+          onChange={event => setVersion(event.target.value)}>
+          {VERSIONS.map(version => (
+            <option key={version} value={version}>
+              {versionToText(version)}
+            </option>
+          ))}
+        </select>
+        {/* Changing versions is a JS only mechanism. To help crawlers find other versions, we add hidden links. */}
+        {VERSIONS.map(version => (
+          <A className="hidden" key={version} href={`/versions/${version}`} />
+        ))}
+      </div>
+    </div>
+  );
+}
+
+function versionToText(version: string): string {
+  if (version === 'unversioned') {
+    return 'Unversioned';
+  } else if (version === 'latest') {
+    return `${versionToText(LATEST_VERSION)} (latest)`;
+  } else if (BETA_VERSION && version === BETA_VERSION.toString()) {
+    return `${versionToText(BETA_VERSION.toString())} (beta)`;
+  }
+  return `SDK ${version.substring(1, 3)}`;
+}
