+++ docs/ui/components/Sidebar/SidebarHead.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { theme, DocsLogo } from '@expo/styleguide';
+import { theme, DocsLogo, LinkBase } from '@expo/styleguide';
 import { spacing } from '@expo/styleguide-base';
 import {
   ArrowLeftIcon,
@@ -9,10 +9,11 @@ import {
   BookOpen02DuotoneIcon,
 } from '@expo/styleguide-icons';
 
+import { ApiVersionSelect } from './ApiVersionSelect';
+
 import { shouldShowFeaturePreviewLink } from '~/constants/FeatureFlags.cjs';
 import { Search } from '~/ui/components/Search';
 import { SidebarSingleEntry } from '~/ui/components/Sidebar/SidebarSingleEntry';
-import { A } from '~/ui/components/Text';
 
 type SidebarHeadProps = {
   sidebarActiveGroup: string;
@@ -21,51 +22,56 @@ type SidebarHeadProps = {
 export const SidebarHead = ({ sidebarActiveGroup }: SidebarHeadProps) => {
   if (sidebarActiveGroup === 'archive') {
     return (
-      <div css={sidebarHeadContainerStyle}>
-        <A isStyled href="/" css={sidebarBackLinkStyle}>
+      <div css={sidebarHeadContainerStyle} className="!p-1.5">
+        <LinkBase
+          href="/"
+          className="flex gap-3 items-center p-2.5 rounded-md text-secondary hocus:bg-element">
           <ArrowLeftIcon className="text-icon-secondary" />
           Back
-        </A>
+        </LinkBase>
       </div>
     );
   }
 
   return (
-    <div css={sidebarHeadContainerStyle}>
-      <Search />
-      <SidebarSingleEntry
-        href="/"
-        title="Home"
-        Icon={Home02DuotoneIcon}
-        isActive={sidebarActiveGroup === 'home'}
-      />
-      <SidebarSingleEntry
-        href="/guides/overview"
-        title="Guides"
-        Icon={BookOpen02DuotoneIcon}
-        isActive={sidebarActiveGroup === 'general'}
-      />
-      <SidebarSingleEntry
-        href="/versions/latest"
-        title="Reference"
-        Icon={DocsLogo}
-        isActive={sidebarActiveGroup === 'reference'}
-      />
-      <SidebarSingleEntry
-        href="/tutorial/introduction/"
-        title="Learn"
-        Icon={GraduationHat02DuotoneIcon}
-        isActive={sidebarActiveGroup === 'learn'}
-      />
-      {shouldShowFeaturePreviewLink() && (
+    <>
+      <div css={sidebarHeadContainerStyle}>
+        <Search />
         <SidebarSingleEntry
-          href="/feature-preview"
-          title="Feature Preview"
-          Icon={Stars02DuotoneIcon}
-          isActive={sidebarActiveGroup === 'featurePreview' || sidebarActiveGroup === 'preview'}
+          href="/"
+          title="Home"
+          Icon={Home02DuotoneIcon}
+          isActive={sidebarActiveGroup === 'home'}
         />
-      )}
-    </div>
+        <SidebarSingleEntry
+          href="/guides/overview"
+          title="Guides"
+          Icon={BookOpen02DuotoneIcon}
+          isActive={sidebarActiveGroup === 'general'}
+        />
+        <SidebarSingleEntry
+          href="/versions/latest"
+          title="Reference"
+          Icon={DocsLogo}
+          isActive={sidebarActiveGroup === 'reference'}
+        />
+        <SidebarSingleEntry
+          href="/tutorial/introduction/"
+          title="Learn"
+          Icon={GraduationHat02DuotoneIcon}
+          isActive={sidebarActiveGroup === 'learn'}
+        />
+        {shouldShowFeaturePreviewLink() && (
+          <SidebarSingleEntry
+            href="/feature-preview"
+            title="Feature Preview"
+            Icon={Stars02DuotoneIcon}
+            isActive={sidebarActiveGroup === 'featurePreview' || sidebarActiveGroup === 'preview'}
+          />
+        )}
+      </div>
+      <ApiVersionSelect />
+    </>
   );
 };
 
@@ -76,11 +82,4 @@ const sidebarHeadContainerStyle = css({
   borderBottom: `1px solid ${theme.border.default}`,
   background: theme.background.default,
   gap: spacing[0.5],
-});
-
-const sidebarBackLinkStyle = css({
-  color: theme.text.secondary,
-  display: 'flex',
-  gap: spacing[3],
-  alignItems: 'center',
 });
